+++ example_applications/web_app/config/config.exs
@@ -50,8 +50,8 @@ config :logger, :console,
   format: "$time $metadata[$level] $message\n",
   metadata: [:request_id]
 
-# Use Jason for JSON parsing in Phoenix
-config :phoenix, :json_library, Jason
+# Use JSON falling back to Jason for JSON parsing in Phoenix
+config :phoenix, :json_library, PromEx.JSON
 
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.